SyntaxHighlight: Support highlighting in reply-indicator
Moved highlightCode to `util/html.js`
